+++ frontend/App.tsx
@@ -5,62 +5,15 @@
  * @format
  */
 
-import React, {useEffect} from 'react';
+import React from 'react';
 import {Platform, StatusBar, useColorScheme, View} from 'react-native';
 import {Colors} from 'react-native/Libraries/NewAppScreen';
 import {PRIMARY_COLOR} from './src/helper/Theme';
-import {NavigationContainer, useNavigation} from '@react-navigation/native';
+import {NavigationContainer} from '@react-navigation/native';
 import StackNavigation from './src/navigations/StackNavigation';
 import {SafeAreaProvider} from 'react-native-safe-area-context';
 import {QueryClient, QueryClientProvider} from '@tanstack/react-query';
-import {BackHandler} from 'react-native';
-import {Alert} from 'react-native';
 
-const BackButtonHandler = () => {
-  const navigation = useNavigation();
-  console.log('Navigation', navigation?.getState());
-  useEffect(() => {
-    const backAction = () => {
-      if (!navigation || !navigation.canGoBack()) {
-        return;
-      }
-      const currentRoute =
-        navigation?.getState()?.routes[navigation?.getState()?.index || 0]
-          ?.name;
-
-      if (
-        currentRoute === 'LoginScreen' ||
-        currentRoute === 'TabNavigation' ||
-        currentRoute === 'Profile' ||
-        currentRoute === 'Home' ||
-        currentRoute === 'Podcasts'
-      ) {
-        Alert.alert('Warning', 'Do you want to exit?', [
-          {text: 'No', onPress: () => null},
-          {
-            text: 'Yes',
-            onPress: () => {
-              BackHandler.exitApp();
-              console.log('State', currentRoute);
-            },
-          },
-        ]);
-        return true; // Prevent default behavior
-      } else {
-        navigation.goBack(); // Allow back navigation for other screens
-      }
-    };
-
-    const backHandler = BackHandler.addEventListener(
-      'hardwareBackPress',
-      backAction,
-    );
-
-    return () => backHandler;
-  }, [navigation]);
-
-  return null; // This component doesn't need to render anything
-};
 const queryClient = new QueryClient();
 function App(): React.JSX.Element {
   //const navigation = useNavigation();
@@ -84,7 +37,6 @@ function App(): React.JSX.Element {
           />
           {
             <NavigationContainer>
-              <BackButtonHandler />
               <StackNavigation />
             </NavigationContainer>
           }
